+++ tests/Stub/CircularDependencyValue1.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Stub;
+
+class CircularDependencyValue1
+{
+    public function __construct(CircularDependencyValue2 $c)
+    {
+        $this->c = $c;
+    }
+}